refactor(account): move interceptor file
